requirements.txt
fastapi==0.115.6
python-decouple==3.8
uvicorn==0.34.0
databases==0.9.0
sqlalchemy==2.0.37
asyncpg==0.29.0
psycopg2-binary==2.9.9